Clarify extraneous logging (#1017)
* style: Clarify unclear warning about junit report

Log the actual error (fail to fetch/parse) with debug level, but now log
a warning to the user of the effect of the failure.

[INT-68]

* style: Unactionable warnings to debug level

If saucectl fails to report to the test-runs api, just log a debug level
message for now. It's a non-recoverable state (e.g. saucectl takes no
further action on failure) and its unclear what effect this would even
have on the user.

* Improve log message